wip: re-work markdown 'how to create...' section into python
